requirements.txt
signalwire-agents==1.0.22
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
uvicorn>=0.34.2
